# cmd/api/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/v2/tenants/{tenant_name}/services/{service_alias}/event-log": {
      "post": {
        "description": "get log by level",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "Get the operation log of the specified operation",
        "operationId": "logByAction",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ServiceAlias",
            "name": "service_alias",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "level",
                "event_id"
              ],
              "properties": {
                "event_id": {
                  "description": "eventID\nin: body",
                  "type": "string",
                  "x-go-name": "EventID"
                },
                "level": {
                  "description": "Log level info/debug/error\nin: body",
                  "type": "string",
                  "x-go-name": "Level"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Uniform return format",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
    "/v2/tenants/{tenant_name}/services/{service_alias}/log": {
      "post": {
        "description": "get last x lines logs",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "Get the latest specified number of logs",
        "operationId": "lastLinesLogs",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ServiceAlias",
            "name": "service_alias",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "lines"
              ],
              "properties": {
                "lines": {
                  "description": "Rows\nin: body",
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Lines"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Uniform return format",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
    "/v2/tenants/{tenant_name}/services/{service_alias}/log-file": {
      "get": {
        "description": "get log list",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "Get application log list",
        "operationId": "logList",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ServiceAlias",
            "name": "service_alias",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "Uniform return format",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
    "/v2/tenants/{tenant_name}/services/{service_alias}/log-file/{file_name}": {
      "get": {
        "description": "get log file",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "Download application specific log",
        "operationId": "logFile",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ServiceAlias",
            "name": "service_alias",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "FileName",
            "name": "file_name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "Uniform return format",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
    "/v2/tenants/{tenant_name}/services/{service_alias}/log-instance": {
      "get": {
        "description": "get log socket",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "Get application log web-socket instance",
        "operationId": "logSocket",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ServiceAlias",
            "name": "service_alias",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "Uniform return format",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Values": {
      "description": "It is typically used for query parameters and form values.\nUnlike in the http.Header map, the keys in a Values map\nare case-sensitive.",
      "type": "object",
      "title": "Values maps a string key to a list of values.",
      "additionalProperties": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "x-go-package": "net/url"
    }
  },
  "responses": {
    "commandResponse": {
      "description": "CommandResponse api unified return structure",
      "schema": {
        "type": "object",
        "properties": {
          "bean": {
            "description": "Single Resource Entity",
            "type": "object",
            "x-go-name": "Bean"
          },
          "list": {
            "description": "Resource List",
            "type": "object",
            "x-go-name": "List"
          },
          "msg": {
            "description": "API error message",
            "type": "string",
            "x-go-name": "Msg"
          },
          "number": {
            "description": "Total number of data sets",
            "type": "integer",
            "format": "int64",
            "x-go-name": "ListAllNumber"
          },
          "page": {
            "description": "Current page number",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Page"
          },
          "validation_error": {
            "$ref": "#/definitions/Values"
          }
        }
      }
    }
  }
}